zod tried
