+++ package.json
@@ -45,7 +45,8 @@
     "react-toastify": "^10.0.5",
     "sass": "^1.77.0",
     "sharp": "^0.33.3",
-    "winston": "^3.13.0"
+    "winston": "^3.13.0",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@pinetwork-js/api-typing": "^0.7.0",
