+++ src/services/api.ts
@@ -1,7 +1,8 @@
 'use server';
 
 import axios from 'axios';
-import { IUser, CreateReviewType, CreateSellerType } from '@/constants/types';
+import { z } from 'zod';
+import { IUser, ReviewFeedbackType, CreateSellerType } from '@/constants/types';
 import axiosClient, { setAuthToken } from "@/config/client";
 
 const API = axios.create({
@@ -123,69 +124,3 @@ export const updateSeller = async (sellerId:string, formData: FormData) => {
   }
 };
 
-// Fetch a single review for a seller
-export const fetchSingleReview = async (reviewID: string) => {
-  try {
-    const response = await API.get(`/review-feedback/single/${reviewID}`);
-    return response.data;
-  } catch (error) {
-    console.error(`Error fetching review with ID ${reviewID}:`, error);
-    throw error;
-  }
-};
-
-// Fetch reviews for a seller
-export const fetchReviews = async (sellerId:string) => {
-  try {
-    const response = await API.get(`/review-feedback/${sellerId}`);
-    return response.data;
-  } catch (error) {
-    console.error(`Error fetching reviews for seller with ID ${sellerId}:`, error);
-    throw error;
-  }
-};
-
-// Create a new review
-export const createReview = async (auth:IUser, props: CreateReviewType, token: string) => {
-  const formData = new FormData();
-
-  formData.append('comment', props.comment);
-  formData.append('rating', props.rating.toString());
-  formData.append('review_receiver_id', props.seller);
-  formData.append('review_giver_id', props.user);
-  props.image.forEach(file => formData.append('images', file));
-  formData.append('reply_to_review_id', props.replyId || '')
-
-  // formData.append('reply_to_review_id', props.replyId)
-
-  console.log('this is form data:', formData)
-  
-  try {
-    const response = await API.post('/review-feedback/add', formData, {
-      headers: {
-        'Content-Type': 'multipart/form-data',
-        'Authorization': `Bearer ${token}`
-      },
-    });
-    return response.data;
-  } catch (error) {
-    console.error('Error creating review:', error);
-    throw error;
-  }
-};
-
-// Update a review
-export const updateReview = async (reviewId:string, formData:FormData, authToken: string) => {
-  try {
-    const response = await API.put(`/review-feedback/${reviewId}`, formData, {
-      headers: {
-        'Content-Type': 'multipart/form-data',
-        'Authorization': `Bearer ${authToken}`,
-      },
-    });
-    return response.data;
-  } catch (error) {
-    console.error(`Error updating review with ID ${reviewId}:`, error);
-    throw error;
-  }
-};
